Update ids in annotations when publishing.
