+++ src/models/document/document-content.ts
@@ -1,5 +1,12 @@
-import { Instance, SnapshotIn } from "mobx-state-tree";
+import { getSnapshot, Instance, SnapshotIn } from "mobx-state-tree";
+import { cloneDeep, each } from "lodash";
+
 import { DocumentContentModelWithAnnotations } from "./document-content-with-annotations";
+import { TileRowSnapshotOutType } from "./tile-row";
+import { isArrowAnnotation } from "../annotations/arrow-annotation";
+import { getTileContentInfo } from "../tiles/tile-content-info";
+import { ITileModelSnapshotOut } from "../tiles/tile-model";
+import { uniqueId } from "../../utilities/js-utils";
 
 /**
  * The DocumentContentModel is the combination of 3 parts:
@@ -23,7 +30,89 @@ import { DocumentContentModelWithAnnotations } from "./document-content-with-ann
  * Note: the name "DocumentContent" is important because it is used in other
  * parts of the code to find a MST parent with this name.
  */
-export const DocumentContentModel = DocumentContentModelWithAnnotations.named("DocumentContent");
+export const DocumentContentModel = DocumentContentModelWithAnnotations.named("DocumentContent")
+.views(self => ({
+  snapshotWithUniqueIds(asTemplate = false) {
+    const snapshot = cloneDeep(getSnapshot(self));
+    const tileIdMap: { [id: string]: string } = {};
+    const annotationIdMap: { [id: string]: string } = {};
+
+    snapshot.tileMap = (tileMap => {
+      const _tileMap: { [id: string]: ITileModelSnapshotOut } = {};
+      each(tileMap, (tile, id) => {
+        tileIdMap[id] = tile.id = uniqueId();
+        _tileMap[tile.id] = tile;
+      });
+      return _tileMap;
+    })(snapshot.tileMap);
+
+    // Update the sharedModels with new tile ids
+    each(snapshot.sharedModelMap, (sharedModelEntry, id) => {
+      const _tiles = cloneDeep(sharedModelEntry.tiles);
+      sharedModelEntry.tiles = [];
+      _tiles.forEach(tile => {
+        sharedModelEntry.tiles.push(tileIdMap[tile]);
+      });
+      // Update references to provider
+      if (sharedModelEntry.provider) {
+        sharedModelEntry.provider = tileIdMap[sharedModelEntry.provider];
+      }
+      const sharedModel = sharedModelEntry.sharedModel;
+      // TODO: Figure out why linting is failing
+      // if ("providerId" in sharedModel && typeof sharedModel.providerId === "string") {
+      //   sharedModel.providerId = tileIdMap[sharedModel.providerId];
+      if ("providerId" in sharedModel) {
+        const providedSharedModel = sharedModel as any;
+        if (typeof providedSharedModel.providerId === "string") {
+          providedSharedModel.providerId = tileIdMap[providedSharedModel.providerId];
+        }
+      }
+    });
+    // TODO: Give the shared models new ids
+
+    // Update annotations with new tile ids
+    each(snapshot.annotations, (annotation, id) => {
+      // TODO Move into functions for specific annotation types
+      if (isArrowAnnotation(annotation)) {
+        if (annotation.sourceObject?.tileId) {
+          annotation.sourceObject.tileId = tileIdMap[annotation.sourceObject.tileId];
+        }
+        if (annotation.targetObject?.tileId) {
+          annotation.targetObject.tileId = tileIdMap[annotation.targetObject.tileId];
+        }
+      }
+      annotationIdMap[annotation.id] = uniqueId();
+      annotation.id = annotationIdMap[annotation.id];
+    });
+
+    each(snapshot.tileMap, tile => {
+      getTileContentInfo(tile.content.type)
+        ?.contentSnapshotPostProcessor?.(tile.content, tileIdMap, asTemplate);
+    });
+
+    snapshot.rowMap = (rowMap => {
+      const _rowMap: { [id: string]: TileRowSnapshotOutType } = {};
+      each(rowMap, (row, id) => {
+        tileIdMap[id] = row.id = uniqueId();
+        row.tiles = row.tiles.map(tileLayout => {
+          tileLayout.tileId = tileIdMap[tileLayout.tileId];
+          return tileLayout;
+        });
+        _rowMap[row.id] = row;
+      });
+      return _rowMap;
+    })(snapshot.rowMap);
+
+    snapshot.rowOrder = snapshot.rowOrder.map(rowId => tileIdMap[rowId]);
+
+    return snapshot;
+  },
+}))
+.views(self => ({
+  publish() {
+    return JSON.stringify(self.snapshotWithUniqueIds());
+  },
+}));
 
 export type DocumentContentModelType = Instance<typeof DocumentContentModel>;
 export type DocumentContentSnapshotType = SnapshotIn<typeof DocumentContentModel>;
